Rearrange tactic folder
