+++ src/software/ai/hl/stp/tactic/BUILD
@@ -25,68 +25,6 @@ cc_test(
     ],
 )
 
-cc_library(
-    name = "chip_tactic",
-    srcs = ["chip_tactic.cpp"],
-    hdrs = [
-        "chip_fsm.h",
-        "chip_tactic.h",
-    ],
-    deps = [
-        ":get_behind_ball_tactic",
-        ":tactic",
-        "//shared:constants",
-        "//software/ai/hl/stp/action:chip_action",
-        "//software/logger",
-    ],
-)
-
-cc_test(
-    name = "chip_fsm_test",
-    srcs = ["chip_fsm_test.cpp"],
-    deps = [
-        ":chip_tactic",
-        "//software/test_util",
-        "@gtest//:gtest_main",
-    ],
-)
-
-cc_test(
-    name = "chip_tactic_test",
-    srcs = ["chip_tactic_test.cpp"],
-    deps = [
-        ":chip_tactic",
-        "//software/test_util",
-        "@gtest//:gtest_main",
-    ],
-)
-
-cc_library(
-    name = "kick_tactic",
-    srcs = ["kick_tactic.cpp"],
-    hdrs = [
-        "kick_fsm.h",
-        "kick_tactic.h",
-    ],
-    deps = [
-        ":get_behind_ball_tactic",
-        ":tactic",
-        "//shared:constants",
-        "//software/ai/hl/stp/action:kick_action",
-        "//software/logger",
-    ],
-)
-
-cc_test(
-    name = "kick_fsm_test",
-    srcs = ["kick_fsm_test.cpp"],
-    deps = [
-        ":kick_tactic",
-        "//software/test_util",
-        "@gtest//:gtest_main",
-    ],
-)
-
 cc_test(
     name = "kickoff_chip_tactic_test",
     srcs = ["kickoff_chip_tactic_test.cpp"],
@@ -112,7 +50,7 @@ cc_library(
     srcs = ["kickoff_chip_tactic.cpp"],
     hdrs = ["kickoff_chip_tactic.h"],
     deps = [
-        ":chip_tactic",
+            "//software/ai/hl/stp/tactic/chip:chip_tactic",
         "//shared:constants",
         "//software/ai/hl/stp/action:chip_action",
         "//software/logger",
@@ -205,85 +143,6 @@ cc_test(
 )
 
 cc_library(
-    name = "move_tactic",
-    srcs = ["move_tactic.cpp"],
-    hdrs = [
-        "move_fsm.h",
-        "move_tactic.h",
-    ],
-    deps = [
-        ":tactic",
-        "//shared:constants",
-        "//software/ai/hl/stp/action:move_action",
-        "//software/ai/intent:move_intent",
-        "//software/logger",
-    ],
-)
-
-cc_test(
-    name = "move_fsm_test",
-    srcs = ["move_fsm_test.cpp"],
-    deps = [
-        ":move_tactic",
-        "//software/test_util",
-        "@gtest//:gtest_main",
-    ],
-)
-
-cc_test(
-    name = "move_tactic_test",
-    srcs = ["move_tactic_test.cpp"],
-    deps = [
-        ":move_tactic",
-        "//software/test_util",
-        "@gtest//:gtest_main",
-    ],
-)
-
-cc_test(
-    name = "simulated_move_tactic_test",
-    srcs = ["simulated_move_tactic_test.cpp"],
-    deps = [
-        "//software/ai/hl/stp/tactic:move_tactic",
-        "//software/simulated_tests:simulated_tactic_test_fixture",
-        "//software/simulated_tests:simulated_test_main",
-        "//software/simulated_tests/validation:validation_function",
-        "//software/simulated_tests/validation_functions",
-        "//software/test_util",
-        "//software/time:duration",
-        "//software/world",
-    ],
-)
-
-cc_library(
-    name = "get_behind_ball_tactic",
-    srcs = ["get_behind_ball_tactic.cpp"],
-    hdrs = [
-        "get_behind_ball_fsm.h",
-        "get_behind_ball_tactic.h",
-    ],
-    deps = [
-        ":tactic",
-        "//shared:constants",
-        "//software/ai/hl/stp/action:move_action",
-        "//software/ai/intent:move_intent",
-        "//software/geom:triangle",
-        "//software/geom/algorithms",
-        "//software/logger",
-    ],
-)
-
-cc_test(
-    name = "get_behind_ball_fsm_test",
-    srcs = ["get_behind_ball_fsm_test.cpp"],
-    deps = [
-        ":get_behind_ball_tactic",
-        "//software/test_util",
-        "@gtest//:gtest_main",
-    ],
-)
-
-cc_library(
     name = "passer_tactic",
     srcs = ["passer_tactic.cpp"],
     hdrs = ["passer_tactic.h"],
@@ -329,7 +188,7 @@ cc_library(
     srcs = ["penalty_setup_tactic.cpp"],
     hdrs = ["penalty_setup_tactic.h"],
     deps = [
-        ":move_tactic",
+        "//software/ai/hl/stp/tactic/move:move_tactic",
         "//shared:constants",
         "//software/ai/evaluation:calc_best_shot",
         "//software/ai/hl/stp/action:kick_action",
@@ -442,56 +301,18 @@ cc_test(
 )
 
 cc_library(
-    name = "stop_tactic",
-    srcs = ["stop_tactic.cpp"],
-    hdrs = [
-        "stop_fsm.h",
-        "stop_tactic.h",
-    ],
-    deps = [
-        ":tactic",
-        "//shared:constants",
-        "//software/ai/hl/stp/action:move_action",
-        "//software/ai/hl/stp/action:stop_action",
-        "//software/ai/intent:move_intent",
-        "//software/ai/intent:stop_intent",
-        "//software/logger",
-    ],
-)
-
-cc_test(
-    name = "stop_fsm_test",
-    srcs = ["stop_fsm_test.cpp"],
-    deps = [
-        ":stop_tactic",
-        "//software/test_util",
-        "@gtest//:gtest_main",
-    ],
-)
-
-cc_test(
-    name = "stop_tactic_test",
-    srcs = ["stop_tactic_test.cpp"],
-    deps = [
-        ":stop_tactic",
-        "//software/test_util",
-        "@gtest//:gtest_main",
-    ],
-)
-
-cc_library(
     name = "all_tactics",
     hdrs = ["all_tactics.h"],
     deps = [
         ":cherry_pick_tactic",
-        ":chip_tactic",
+        "//software/ai/hl/stp/tactic/chip:chip_tactic",
         ":crease_defender_tactic",
         ":defense_shadow_enemy_tactic",
-        ":get_behind_ball_tactic",
+        "//software/ai/hl/stp/tactic/get_behind_ball:get_behind_ball_tactic",
         ":goalie_tactic",
-        ":kick_tactic",
+        "//software/ai/hl/stp/tactic/kick:kick_tactic",
         ":kickoff_chip_tactic",
-        ":move_tactic",
+        "//software/ai/hl/stp/tactic/move:move_tactic",
         ":passer_tactic",
         ":penalty_kick_tactic",
         ":penalty_setup_tactic",
@@ -499,7 +320,7 @@ cc_library(
         ":shadow_enemy_tactic",
         ":shadow_free_kicker_tactic",
         ":shoot_goal_tactic",
-        ":stop_tactic",
+        "//software/ai/hl/stp/tactic/stop:stop_tactic",
         ":tactic",
     ],
 )
