+++ src/software/ai/hl/stp/tactic/chip/chip_fsm.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "software/ai/hl/stp/tactic/get_behind_ball_fsm.h"
+#include "software/ai/hl/stp/tactic/get_behind_ball/get_behind_ball_fsm.h"
 #include "software/ai/hl/stp/tactic/tactic.h"
 #include "software/ai/intent/chip_intent.h"
