+++ src/software/ai/hl/stp/tactic/chip/chip_tactic.h
@@ -2,7 +2,7 @@
 
 #include <queue>
 
-#include "software/ai/hl/stp/tactic/chip_fsm.h"
+#include "software/ai/hl/stp/tactic/chip/chip_fsm.h"
 #include "software/ai/hl/stp/tactic/tactic.h"
 #include "software/ai/intent/chip_intent.h"
