+++ src/software/ai/hl/stp/tactic/chip/chip_tactic_test.cpp
@@ -1,4 +1,4 @@
-#include "software/ai/hl/stp/tactic/chip_tactic.h"
+#include "software/ai/hl/stp/tactic/chip/chip_tactic.h"
 
 #include <gtest/gtest.h>
