+++ src/software/ai/hl/stp/tactic/move/BUILD
@@ -1,3 +1,5 @@
+package(default_visibility = ["//visibility:public"])
+
 cc_library(
     name = "move_tactic",
     srcs = ["move_tactic.cpp"],
