+++ src/software/ai/hl/stp/tactic/stop/stop_tactic.h
@@ -1,39 +1,32 @@
 #pragma once
 
 #include "software/ai/hl/stp/action/move_action.h"  // TODO (#1888): remove this dependency
-#include "software/ai/hl/stp/tactic/get_behind_ball_fsm.h"
+#include "software/ai/hl/stp/tactic/stop/stop_fsm.h"
 #include "software/ai/hl/stp/tactic/tactic.h"
+#include "software/ai/intent/stop_intent.h"
 
 /**
- * The GetBehindBallTactic will move the assigned robot to the given destination and
- * arrive with the specified final orientation and speed
+ * The StopTactic will stop the robot from moving. The robot will actively try and brake
+ * to come to a halt unless is it told to coast, in which case it will coast to a stop.
  */
-class GetBehindBallTactic : public Tactic
+class StopTactic : public Tactic
 {
    public:
     /**
-     * Creates a new GetBehindBallTactic
-     *
+     * Creates a new StopTactic
+     t
      * @param loop_forever Whether or not this Tactic should never complete. If true, the
      * tactic will be restarted every time it completes
      */
-    explicit GetBehindBallTactic(bool loop_forever);
+    explicit StopTactic(bool coast);
 
-    GetBehindBallTactic() = delete;
+    StopTactic() = delete;
 
     void updateWorldParams(const World& world) override;
 
     /**
-     * Updates the control parameters for this GetBehindBallTactic.
-     *
-     * @param ball_location The location of the ball when it will be chipped or kicked
-     * @param chick_direction The direction to kick or chip
-     */
-    void updateControlParams(const Point& ball_location, Angle chick_direction);
-
-    /**
-     * Calculates the cost of assigning the given robot to this Tactic. Prefers robots
-     * closer to the destination
+     * Calculates the cost of assigning the given robot to this Tactic. Prefers all robots
+     * equally
      *
      * @param robot The robot to evaluate the cost for
      * @param world The state of the world with which to perform the evaluation
@@ -47,9 +40,11 @@ class GetBehindBallTactic : public Tactic
 
    private:
     void calculateNextAction(ActionCoroutine::push_type& yield) override;
+
     void updateIntent(const TacticUpdate& tactic_update) override;
 
-    BaseFSM<GetBehindBallFSM> fsm;
+    BaseFSM<StopFSM> fsm;
 
-    GetBehindBallFSM::ControlParams control_params;
+    // Whether or not the robot should coast to a stop
+    bool coast;
 };
