+++ src/software/ai/hl/stp/tactic/stop/stop_tactic_test.cpp
@@ -1,4 +1,4 @@
-#include "software/ai/hl/stp/tactic/stop_tactic.h"
+#include "software/ai/hl/stp/tactic/stop/stop_tactic.h"
 
 #include <gtest/gtest.h>
